+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.11] - 2022-01-23
+
+### Added
+
+- `updateUserEmail` method, to update user's email address
+- `deleteUser` method, to delete user's account
+
 ## [0.9.10] - 2022-01-11
 
 ### Added